% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computeAdvection}
\alias{computeAdvection}
\title{Compute advective flux in g/(m^3 * min)}
\usage{
computeAdvection(inputDOFlux, DOconc, data)
}
\arguments{
\item{inputDOFlux}{dissolved oxygen mass flux of input water (g/\link{m^3 * min})}

\item{DOconc}{dissolved oxygen concentration (g/m^3)}

\item{data}{Fixed site data, including at least Q, area, and dx}
}
\value{
dissolved oxygen concentration flux (g/\link{m^3 * min})
}
\description{
Compute advective flux in g/(m^3 * min)
}
