% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computeGPP}
\alias{computeGPP}
\title{Compute gross primary productivity from light (g O2 / \link{m^2 * day})}
\usage{
computeGPP(PAR, lP1, lP2)
}
\arguments{
\item{lP1}{The log of the slope of the PI curve}

\item{lP2}{The log of the saturation term of the PI curve}

\item{data}{Fixed site data, including at least Q, area, and dx}
}
\value{
double; computed GPP
}
\description{
Compute gross primary productivity from light (g O2 / \link{m^2 * day})
}
\details{
P1 is in units of (W*day)/gO2, P2 is in (m^2 * day) / gO2
}
\references{
Uehlinger U. et al. 2000. Variability of photosynthesis-irradiance curves and
ecosystem respiration in a small river. Freshwater Biology 44:493-507.
}
