% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computeInputDOFlux}
\alias{computeInputDOFlux}
\title{Computes the input dissolved oxygen flux}
\usage{
computeInputDOFlux(upQ, upDO, latQ, latDO)
}
\arguments{
\item{upQ}{Upstream discharge (m^3/s)}

\item{upDO}{Upstream dissolved oxygen concentration (g/m^3)}

\item{latQ}{Lateral input discharge (m^3/s)}

\item{latDO}{Lateral input DO concentration (g/m^3)}
}
\value{
double; the input flux (mg/min) of dissolved oxygen
}
\description{
Computes the input dissolved oxygen flux
}
