% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computeRF}
\alias{computeRF}
\title{Computes reaeration flux (g/(m^2 * day))}
\usage{
computeRF(temp, pressure, DO, k600)
}
\arguments{
\item{temp}{Water temperature (degrees C)}

\item{pressure}{Atmospheric pressure (hPa)}

\item{DO}{dissolved oxygen concentration, mg/L}

\item{k600}{Gas transfer coefficient for Schmidt number of 600}
}
\value{
computed rearation flux
}
\description{
Computes reaeration flux (g/(m^2 * day))
}
