% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onestation.r
\name{DO_logprob}
\alias{DO_logprob}
\title{Log posterior probability for the DO curve given empirical data}
\usage{
DO_logprob()
}
\description{
Log posterior probability for the DO curve given empirical data
}
