% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onestation.r
\name{DO_predict}
\alias{DO_predict}
\title{Predict a time series of dissolved oxygen concentration}
\usage{
DO_predict(initial, times, params, data, dt = 1, method = c("euler",
  "lsoda"))
}
\arguments{
\item{initial}{initial dissolved oxygen concentration}

\item{times}{The times at which to solve the system}

\item{params}{a list of parameters; see details}

\item{data}{a list of data for the ode; see details}

\item{dt}{Time step for integration; only used if \code{method == 'euler'}}

\item{method}{The integration method to use; default is euler}
}
\value{
Time series of dissolved oxygen concentrations
}
\description{
Predict a time series of dissolved oxygen concentration
}
\details{
Light and temperature time series will be approximated using linear interpolation
at the desired time steps

\code{params} must be a list including the following named items:
\itemize{
\item \code{P1} $(W min g^{-1} O_2)$ inverse of the slope of a photosynthesis–irradiance curve at low
light intensity
\item \code{P2} $(m^2 min g^{-1} O_2)$ inverse maximum photosynthesis rate; can be zero to assume
GPP is linear with light intensity instead of saturating
\item \code{k600} coefficient of gas exchange for a gas with a Schmidt number of 600; not supported,
currently fixed to 2/(24*60) (see metabolism_functions.r/kT())
\item \code{ER24_20} daily ecosystem respiration rate, standardized at 20 degrees C
}

\code{data} is a named list of (constant) data items, including the following:
\itemize{
\item \code{PAR}  2-column data frame; first column is light, second is time of observation
\item \code{temp} 2-column data frame; first column is temperature, second is time of observation
\item \code{P}    pressure, in atmospheres
\item \code{z}    Depth, in meters
}
}
