% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.r
\name{DOCalibration}
\alias{DOCalibration}
\title{Calibrate a dissolved oxygen model}
\usage{
DOCalibration(initial, data, model = c("onestation", "twostation",
  "nstation"), method = c("laplace", "stan"), nsamples = 1000,
  prior = list(), dt = 1, ...)
}
\arguments{
\item{initial}{a named vector with one element per parameter giving the initial values}

\item{data}{Dataset to be passed along to the model chosen}

\item{model}{The model to use; see 'details'}

\item{method}{Calibration method; either \code{laplace} approximation or \code{mcmc} (not implemented)}

\item{nsamples}{The number of posterior samples to return}

\item{prior}{Priors for the function chosen}

\item{...}{Additional arguments to pass to the sampler}
}
\value{
A matrix of posterior samples of all parameters
}
\description{
Calibrate a dissolved oxygen model
}
\details{
The \code{model} parameter selects which model will be used; one of:
\itemize{
\item onestation: The one-station model from Fuß et al 2017; params/data from
\code{\link[=oneStation_DOlogprob]{oneStation_DOlogprob()}}
\item twostation: Not implemented
\item nstation: Not implemented
}
}
\references{
Fuß, T. et al. (2017). Land use controls stream ecosystem metabolism by shifting
dissolved organic matter and nutrient regimes. \emph{Freshw Biol} \strong{62}:582–599.
}
