% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{.check_pressure_args}
\alias{.check_pressure_args}
\title{Validate input data for the air pressure function}
\usage{
.check_pressure_args(pressure, stations, out_sites, out_times)
}
\description{
Validate input data for the air pressure function
}
\keyword{internal}
