% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.r
\name{.k_generate_pars}
\alias{.k_generate_pars}
\title{generate parameter sets (assuming independence) for k sims}
\usage{
.k_generate_pars(par, se, n)
}
\description{
generate parameter sets (assuming independence) for k sims
}
\keyword{internal}
