% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.r
\name{.k_sd_sim}
\alias{.k_sd_sim}
\title{generate standard dev of k for a given set of stream variables}
\usage{
.k_sd_sim(S, V, D, Q, pars, eqn)
}
\description{
generate standard dev of k for a given set of stream variables
}
\keyword{internal}
