% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{.osm_summarize_single}
\alias{.osm_summarize_single}
\title{Summarize a single onestation fit}
\usage{
.osm_summarize_single(fit, name, diss_o)
}
\arguments{
\item{fit}{A onestation fit}

\item{name}{Model name}

\item{diss_o}{Optional, DO time series used to fit the model}
}
\description{
Summarize a single onestation fit
}
\keyword{internal}
