% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{.pr_site}
\alias{.pr_site}
\title{get a vector of pressure for a single site}
\usage{
.pr_site(pr, st, si, times, stn_id, site_id, ...)
}
\description{
get a vector of pressure for a single site
}
\keyword{internal}
