% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.r
\name{gpp}
\alias{gpp}
\title{Compute Gross Primary Productivity}
\usage{
gpp(par, P1, P2)
}
\arguments{
\item{par}{light intensity ($W/m^2$)}

\item{P1}{$(W min g^{-1} O_2)$ inverse of the slope of a photosynthesis–irradiance curve at low
light intensity}

\item{P2}{$(m^2 min g^{-1} O_2)$ inverse maximum photosynthesis rate; can be zero to assume
GPP is linear with light intensity instead of saturating}
}
\value{
Photosynthetic rate
}
\description{
Compute Gross Primary Productivity
}
\references{
Uehlinger, U., et al. 2000. Variability of photosynthesis‐irradiance curves and
ecosystem respiration in a small river. \emph{Freshw Biol} \strong{44}:493–507.
}
