% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolism_functions.r
\name{hPaToAtm}
\alias{hPaToAtm}
\title{Convert hPa to atmospheres}
\usage{
hPaToAtm(P)
}
\arguments{
\item{P}{Vector of pressures (in hPa)}
}
\value{
Pressure in atm
}
\description{
Convert hPa to atmospheres
}
