% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{idw_matrix}
\alias{idw_matrix}
\title{Inverse distance weighted interpolation}
\usage{
idw_matrix(vals, dist, pow = 2)
}
\arguments{
\item{vals}{A location-by-value matrix}

\item{dist}{Distance to each value}

\item{pow}{Power to use for distance}
}
\value{
The inverse distance weighted average of vals
}
\description{
Inverse distance weighted interpolation
}
\details{
Interpolation is done assuming locations are in the rows of vals. Thus, if vals
is an n by m matrix, each of the m output values will be the inverse distance weighted
average for the n sites
}
