% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{idw_river}
\alias{idw_river}
\title{Perform inverse distance weighting with a correction for discharge}
\usage{
idw_river(vals, dist, nbQ, Q)
}
\arguments{
\item{vals}{input values of variable to interpolate}

\item{dist}{distance from focal point to vals; negative for downstream sites}

\item{nbQ}{discharge of measured sites}

\item{Q}{discharge of focal site}
}
\value{
interpolated value of vals
}
\description{
This function works on a vector of k output sites, where the length
of each input list and the input vector Q must be k.
For each input site, the corresponding input lists contain vectors of neighbor sites;
so for site k with discharge Q\link{k}, vals[\link{k}] is a vector for the n neighbors of site k.
}
