% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{inSituER}
\alias{inSituER}
\title{Compute ecosystem respiration at in-situ temperature (gO2 / \link{m^2 * day})}
\usage{
inSituER(temperature, ER24_20)
}
\arguments{
\item{temperature}{Water temperature (degrees C)}

\item{ER24_20}{Ecosystem respiration rate at 20 degrees (gO2 / \link{m^2 * day})}
}
\value{
Temperature-adjusted ER
}
\description{
Compute ecosystem respiration at in-situ temperature (gO2 / \link{m^2 * day})
}
