% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_prior.r
\name{k_prior_sim}
\alias{k_prior_sim}
\title{Simulation to develop a prior distribution for k600}
\usage{
k_prior_sim()
}
\value{
A list with parameters for the mean and standard deviation of the prior distribution
for k.
}
\description{
Simulation to develop a prior distribution for k600
}
\details{
This function is retained for historical purposes, it was used to derive the
variance for the informative prior for k used in the model. Parameters from Raymond et al 2012
equation 3.

The prior for k is defined as:
\deqn{\mu = a \times S^{b_s} \times V^{b_v}}{}
\deqn{\sigma = a_{se} + b_{sse} \times S + b_{vse} \times V + b_{svse} \times S \times V}
where S is the slope (unitless, rise over run) and V is the velocity (m/s).
}
\references{
Raymond, P.A. et al. (2012). Scaling the gas transfer velocity and hydraulic
geometry in streams and small rivers. Limnol. Oceanogr., 2, 41–53.
}
\keyword{internal}
