% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kT}
\alias{kT}
\title{Compute gas transfer velocity for oxygen at a given temperature (m/day)}
\usage{
kT(temp, k600)
}
\arguments{
\item{temp}{Water temperature (degrees C)}

\item{k600}{Gas transfer coefficient for Schmidt number of 600 (m/day)}
}
\value{
k at the given temperature (m/day)
}
\description{
Compute gas transfer velocity for oxygen at a given temperature (m/day)
}
\references{
Wanninkhof R. (1992). Relationship between wind speed and gas exchange over the
ocean. Journal of Geophysical Research, 97, 7373.

Van de Bogert, M.C., Carpenter, S.R., Cole, J.J. & Pace, M.L. (2007). Assessing pelagic
and benthic metabolism using free water measurements. Limnology and Oceanography:
Methods, 5, 145–155.
}
