% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onestation.r
\name{oneStation_DOlogprob}
\alias{oneStation_DOlogprob}
\title{Log posterior probability for the DO curve given empirical data}
\usage{
oneStation_DOlogprob(params, data, prior = list(), ...)
}
\arguments{
\item{params}{A named vector of model parameters; see details}

\item{data}{A list of model data, see details}

\item{prior}{A list of prior hyperparameters; see details}

\item{...}{Additional parameters to be passed to \code{\link[=DO_predict]{DO_predict()}}}
}
\value{
The unnormalized log probability of the model given the data
}
\description{
Log posterior probability for the DO curve given empirical data
}
\details{
#' \code{params} must be a vector including the following named items:
\itemize{
\item \code{logP1} $(W min g^{-1} O_2)$ log(inverse of the slope of a photosynthesis–irradiance curve)
at low light intensity
\item \code{logP2} $(m^2 min g^{-1} O_2)$ log(inverse maximum photosynthesis rate)
\item \code{k600} coefficient of gas exchange for a gas with a Schmidt number of 600; not supported,
currently fixed to 2/(24*60) (see \code{\link[=kT]{kT()}})
\item \code{logMinusER24_20} log(-1 * daily ecosystem respiration rate), standardized at 20 degrees C
\item \code{logsd} log(Error standard deviation)
}

\code{data} is a named list of (constant) data items, including the following:
\itemize{
\item \code{DO} 2-column data frame; first column is dissolved oxygen, second is time of observation
\item \code{PAR}  2-column data frame; first column is light, second is time of observation
\item \code{temp} 2-column data frame; first column is temperature, second is time of observation
\item \code{P}    pressure, in atmospheres
\item \code{z}    Depth, in meters
}

\code{prior} is a list of hyperparameters for the priors on parameters listed under \code{params}.
All priors are normal. Thus, each list item should be named (following the names from
\code{params}), and each item should be a vector of length two giving the mean and standard
deviation of the prior for that parameter. Note that each parameter is optional
(as is the entire list); missing items will get a minimally informative normal(0,10)
distribution.
}
