% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onestation.r
\name{oneStation_DOSim}
\alias{oneStation_DOSim}
\title{Produce posterior simulations for DO curve and GPP}
\usage{
oneStation_DOSim(params, data)
}
\arguments{
\item{params}{Parameter posterior matrix returned (e.g.) from \code{\link[=DOCalibration]{DOCalibration()}}.}

\item{data}{Data list as required by \code{\link[=oneStation_DOPredict]{oneStation_DOPredict()}}.}
}
\value{
list including predicted DO and GPP at each time step as well as the prediction times.
}
\description{
Produce posterior simulations for DO curve and GPP
}
