% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onestation.r
\name{oneStation_parTransform}
\alias{oneStation_parTransform}
\title{Transfornm parameters for the one station model}
\usage{
oneStation_parTransform(params, reverse = FALSE)
}
\arguments{
\item{params}{The vector of parameters, see \code{\link[=oneStation_DOlogprob]{oneStation_DOlogprob()}}, or a matrix of
samples (in which case the transformation is applied to app samples)}

\item{reverse}{Should the reverse transformation be applied?}
}
\value{
The transformed parameters
}
\description{
Transfornm parameters for the one station model
}
\details{
This function transforms the parameters of the one station model from the
scale at which the parameters are calibrated to the scale of modelling. If
\code{reverse == TRUE} the opposite transformation is performed.
}
\keyword{internal}
