% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{osat}
\alias{osat}
\title{Compute dissolved oxygen saturation given temperature and pressure (mg/L OR g/m^3)}
\usage{
osat(temp, P)
}
\arguments{
\item{temp}{Water temperature (degrees C)}

\item{P}{atmospheric pressure, in hPa}
}
\value{
oxygen saturation concentration at given temperature and pressure
}
\description{
Compute dissolved oxygen saturation given temperature and pressure (mg/L OR g/m^3)
}
\references{
Benson BB and Krause D. 1984. The concentration and isotopic fractionation of
oxygen dissolved in freshwater and seawater in equilibrium with the atmosphere.
Limnol. Oceanogr., 29, 620–632.

Benson BB and Krause D. 1980. The concentration and isotopic fractionation of gases
dissolved in freshwater in equilibrium with the atmosphere. 1. Oxygen.  Limnol.
Oceanogr., 25(4):662-671
}
