% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{osm_summarize}
\alias{osm_summarize}
\title{Produce a summary from a list of one-station fits}
\usage{
osm_summarize(fits, files, diss_o)
}
\arguments{
\item{fits}{A list of onestation fits, required if \code{files} is not specified}

\item{files}{A vector or list of filenames with .rds files of onestation fits, required if
\code{fits} is not specified}

\item{diss_o}{Optional, a list of vectors with one element per element in fits and files.
Contains the dissolved oxygen time series used to fit each model.}
}
\description{
If many fits are involved, memory usage can be extreme when providing \code{fits} instead of
\code{files.} It is recommended instead to save fits one at a time
}
\details{
If \code{diss_o} is provided, then the model root mean square error and the pearson correlation
(predicted vs calibration) will also be included
}
