% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pixelMetabolism}
\alias{pixelMetabolism}
\title{Compute metabolism using pixels}
\usage{
pixelMetabolism(pixdf, light, temperature, pressure, do_init, lP1, lP2,
  er24_20, k600, variable)
}
\arguments{
\item{pixdf}{The pixel data frame to construct the pixels}

\item{light}{Matrix of light readings; pixels in rows, light readings in columns}

\item{temperature}{Water temperature (degrees C)}

\item{lP1}{The log of the slope of the PI curve}

\item{lP2}{The log of the saturation term of the PI curve}

\item{er24_20}{Ecosystem respiration rate at 20 degrees (gO2 / \link{m^2 * day})}

\item{variable;}{the variable to compute}
}
\value{
Matrix; computed GPP or ER for each pixel at each time step
}
\description{
Compute metabolism using pixels
}
