% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pressureCorrection}
\alias{pressureCorrection}
\title{Compute pressure at a given elevation from a calibration point}
\usage{
pressureCorrection(P, elev, newElev)
}
\arguments{
\item{P}{vector of pressure observations (e.g., a time series)}

\item{elev}{Vector of elevations (in meters)}

\item{calibE}{Elevaiton of calibration point}
}
\value{
Pressure in hPa; in the form of a matrix with elevation as rows and pressure as columns
}
\description{
Compute pressure at a given elevation from a calibration point
}
