% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.r
\name{pressureCorrection_r}
\alias{pressureCorrection_r}
\title{Compute pressure at a given elevation from a calibration point}
\usage{
pressureCorrection_r(elev, calibE, calibP, units = c("hPa", "Pa"))
}
\arguments{
\item{elev}{Vector of elevations (in meters)}

\item{calibE}{Elevaiton of calibration point}

\item{calibP}{vector of pressure observations (e.g., a time series)}
}
\value{
Pressure in hPa; in the form of a matrix with elevation as rows and pressure as columns
}
\description{
Compute pressure at a given elevation from a calibration point
}
