% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolism_functions.r
\name{pressureFromElevation}
\alias{pressureFromElevation}
\title{Compute pressure (in hPa) from elevation}
\usage{
pressureFromElevation(elev)
}
\arguments{
\item{elev}{Vector of elevations}
}
\value{
Pressure for given elevations in hPa
}
\description{
Compute pressure (in hPa) from elevation
}
