% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loggers.r
\name{read_hobo}
\alias{read_hobo}
\title{Read hobo light/temperature logger data}
\usage{
read_hobo(fname, pattern, tformat = "dmyHMS")
}
\arguments{
\item{fname}{file or directory name, see 'details'}

\item{pattern}{grep-compatible search pattern for the filename if fname is a directory}

\item{tformat}{A date/time format, following \link[lubridate:parse_date_time]{lubridate::parse_date_time}}
}
\value{
A \link[data.table:data.table]{data.table::data.table}
}
\description{
Read hobo light/temperature logger data
}
\details{
If \code{fname} is a directory, then \code{pattern} is also required; this allows
choosing the hobo by name
}
