% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loggers.r
\name{read_minidot}
\alias{read_minidot}
\title{Get data from a directory of minidot files in txt (csv) format}
\usage{
read_minidot(fname, skip = 3, ...)
}
\arguments{
\item{fname}{Directory name containing minidot files to read}

\item{skip}{How many header (non-data) lines will be at the beginning of the files? Normally
the default of 3 is fine.}

\item{...}{Additional named arguments to be passed to \code{\link[data.table:fread]{data.table::fread()}} or \code{\link[=read.csv]{read.csv()}}}
}
\value{
A \link[data.table:data.table]{data.table::data.table} or \code{data.frame}
}
\description{
Get data from a directory of minidot files in txt (csv) format
}
\details{
This function searches a directory for minidot data files, reads them in and returns
a data.table or data.frame; all files found in the directory will be collected into a
single table.
}
