% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/light.r
\name{setup_irradiance}
\alias{setup_irradiance}
\title{Create rasters needed for irradiance analysis}
\usage{
setup_irradiance(dem_name, slope_name, aspect_name, lon_name, horizon_pars)
}
\arguments{
\item{dem_name}{character; this parameter (and similar name parameters) tell GRASS what names to
use for the rasters needed for the analysis; they must exist in the GRASS location/mapset if
\code{use_existing == TRUE}.}

\item{horizon_pars}{List of arguments for GRASS74's r.horizon; see the GRASS GIS help files if you wish to change the defaults.}
}
\description{
Create rasters needed for irradiance analysis
}
\keyword{internal}
