% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twostation.r
\name{twoStation_DOPredict}
\alias{twoStation_DOPredict}
\title{Predict a time series of dissolved oxygen concentration}
\usage{
twoStation_DOPredict(ws, initial, times, params, data, dt = 1,
  method = c("euler", "lsoda"))
}
\arguments{
\item{ws}{A \code{\link[WatershedTools:Watershed]{WatershedTools::Watershed()}} object, describing ONLY the two sites}

\item{initial}{initial dissolved oxygen concentration in each station}

\item{times}{The times at which to solve the system}

\item{params}{a named vector of parameters; see details}

\item{data}{a list of data for the ode; see details}

\item{dt}{Time step for integration; only used if \code{method == 'euler'}}

\item{method}{The integration method to use; default is euler}

\item{lateral}{Vector  of length two giving the DO concentration of lateral input
(useful) if discharge varies significantly between the sites, defaults to 0}
}
\value{
Time series of dissolved oxygen concentrations
}
\description{
Predict a time series of dissolved oxygen concentration
}
\details{
Light and temperature time series will be approximated using linear interpolation
at the desired time steps

\code{params} must be a vector. The indices of the vector must map to model parameters
as follows; (where parameters with length two are fit per site):
\itemize{
\item \link{1:2} -- \code{P1} $(W min g^{-1} O_2)$ -- inverse of the slope of a photosynthesis–irradiance
curve at low light intensity
\item \link{3:4} -- \code{P2} $(m^2 min g^{-1} O_2)$ -- inverse maximum photosynthesis rate; can be zero
to assume GPP is linear with light intensity instead of saturating
\item \link{5:6} -- \code{k600} -- coefficient of gas exchange for a gas with a Schmidt number of 600
\item \link{7:8} -- \code{ER24_20} -- daily ecosystem respiration rate, standardized at 20 degrees C
}

\code{data} is a list of length 2 (one per site); each item is a list of (constant) data items,
including the following:
\itemize{
\item \code{PAR}  2-column data frame; first column is light, second is time of observation
\item \code{temp} 2-column data frame; first column is temperature, second is time of observation
\item \code{P}    pressure, in atmospheres
\item \code{z}    Depth, in meters
}
}
