% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolism_functions.r
\name{kT}
\alias{kT}
\title{Compute the temperature-corrected reaction coefficient}
\usage{
kT(temp, k600 = 2/(24 * 60))
}
\arguments{
\item{temp}{Temperature in degrees C}

\item{K600}{gas exchange coefficient}
}
\value{
vector of kT values
}
\description{
Compute the temperature-corrected reaction coefficient
}
\references{
Van de Bogert MC et al. 2007. Assessing pelagic and benthic metabolism using free
water measurements. \emph{Limnol. Oceanogr.: Methods} \strong{5}:145–155. \
Wanninkhof, R. 1992. Relationship between wind speed and gas exchange over the ocean.
\emph{Journal of Geophysical Research} \strong{97}:7373.
}
\keyword{internal}
