% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolism_functions.r
\name{osat}
\alias{osat}
\title{Compute dissolved oxygen saturation given temperature and pressure}
\usage{
osat(temp, P)
}
\arguments{
\item{temp}{Water temperature, in degrees C}

\item{P}{pressure, in atmospheres}
}
\value{
vector of DO saturation values
}
\description{
Compute dissolved oxygen saturation given temperature and pressure
}
\references{
Benson BB and Krause D, Jr. 1980. The concentration and isotopic fractionation of
gases dissolved in freshwater in equilibrium with the atmosphere. 1. Oxygen. \emph{Limnol.
Oceanogr.} \strong{25}(4):662-671. \
Benson, B.B. and Krause, D. 1984. The concentration and isotopic fractionation of
oxygen dissolved in freshwater and seawater in equilibrium with the atmosphere.
\emph{Limnol. Oceanogr.}, \strong{29}:620–632.
}
\keyword{internal}
