% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolism_functions.r
\name{pressureFromElevation}
\alias{pressureFromElevation}
\title{Compute pressure in atmospheres elevation}
\usage{
pressureFromElevation(elev)
}
\arguments{
\item{elev}{Vector of elevations}
}
\value{
Pressure for given elevations
}
\description{
Compute pressure in atmospheres elevation
}
