% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loggers.r
\name{read_hobo}
\alias{read_hobo}
\title{Read hobo light/temperature logger data}
\usage{
read_hobo(fname, pattern)
}
\arguments{
\item{fname}{file or directory name, see 'details'}

\item{pattern}{grep-compatible search pattern for the filename if fname is a directory}
}
\value{
A \link[data.table:data.table]{data.table::data.table} or \code{data.frame}
}
\description{
Read hobo light/temperature logger data
}
\details{
If \code{fname} is a directory, then \code{pattern} is also required; this allows
choosing the hobo by name
}
